requirements.txt
nautilus_trader[ib]==1.220.0  # Supports BarSpecification.from_str
pandas>=2.0.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
PyYAML>=6.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.10.0
